fix(metadata): address review feedback on decorator v2 foundations
