+++ packages/metadata/src/adapters/decorator/v2/index.ts
@@ -6,14 +6,8 @@
  */
 
 export * from './constants';
-export * from './context';
-export * from './draft';
-export * from './handler';
-export * from './helpers';
-export * from './loader';
-export * from './match';
-export * from './preset';
-export * from './registry';
-export * from './source';
+export * from './module';
+export * from './types';
 export * from './typescript';
+export * from './utils';
 export * from './validation';
