+++ packages/metadata/src/adapters/decorator/v2/typescript/index.ts
@@ -5,6 +5,6 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-export * from './argument';
-export * from './decorator-source';
-export * from './jsdoc-source';
+export * from './module';
+export * from './types';
+export * from './utils';
