+++ packages/metadata/src/adapters/decorator/v2/validation/index.ts
@@ -6,4 +6,3 @@
  */
 
 export * from './module';
-export * from './schemas';
